pyproject.toml
[project]
name = "teste-minsait"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dask>=2025.2.0",
    "delta-spark>=3.3.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pyspark>=3.5.5",
]
